Instantiation syntax replaces constructors
Before, when declaring a datatype such as:

    data Bool : Type
    | true  : Bool
    | false : Bool

To create an instance of type Bool, we could use the dot notation to
access one of its constructors. For example:

    Bool.true

Is the first constructor of that datatype. To compile Formality to SIC,
Scott-encodings are used. That means the term above was translated as:

    λt. λf. t

Problem is, in order to make best use of Formality's fusion
capabilities, that isn't enough. For example, when it comes to
Scott-encodings, there are two ways to implement `not`:

    not_A = λbool. (bool (λt. λf. f) (λt. λf. t))
    not_B = λbool. λt. λf. (bool f t)

Both implementations are equivalent, but the second one fuses when
applied repeatedly. The reason is that it shares the "header" of the
returned value (`λt. λf.`) on both branches of the computation. This is
extremelly important, as highlighted on the following post:

https://medium.com/@maiavictor/solving-the-mystery-behind-abstract-algorithms-magical-optimizations-144225164b07

Using the previous compile strategy, we could only express the first
implementation of not:

    let not(bool : Bool) =>
        case bool
        | true  => Bool.false
        | false => Bool.true

Note that `Bool.false` is compiled to `λt. λf. f`, and `Bool.true` is
compiled to `λt. λf. t`, so the whole term is compiled to:
`λbool. (bool (λt. λf. f) (λt. λf. t))`, which is the first definition
of `not`. There was no way to write a `not` term that compiles to the
second definition, which means we can't make use of SIC's runtime fusion.

To solve that, I'm replacing direct constructors by an instantiation
syntax. With it, this:

    MyType.constructor

Becomes this:

    MyType{constructor}

The difference is that, now, we're able to instruct the compiler to place
the header of a Scott-encoding in a different place than its body, and
to share it if needed. That allows us to express, for example, both versions
of not:

    let not_A(bool : Bool) =>
        case bool
        | true  => Bool{false}
        | false => Bool{true}

    let not_B(bool : Bool) =>
        Bool{
            case bool
            | true  => false
            | false => true
        }

Which compile to `not_A` and `not_B` as described previously,
respectivelly. To be more specific, this is how both are compiled,
line to line:

not_A(bool : Bool) =>      -- λbool.
    case bool              -- (bool
    | true  => Bool{false} --   (λt. λf. f)
    | false => Bool{true}  --   (λt. λf. t))

not_B(bool : Bool) =>      -- λbool.
    Bool{                  -- λt. λf.
        case bool          -- (bool
        | true  => false   --   f
        | false => true    --   t)
    }                      --

Note: the syntax could, alternativelly, use the `new` keyword:

    Nat{succ(Nat{succ(Nat{zero})})}

Would be instead:

    new Nat: succ(new Nat: succ(new Nat: zero))

I'm not decided yet.