IDT parameters
This allows specifying Agda-like parameters on IDTs.
For example, the Pair type, previously, was defined/used as:

    data Pair : (A : Type, B : Type) -> Type
    | new : (A : Type, B : Type, x : A, y : B) -> Pair

    let my_pair
        Pair{new(Bool, Bool, Bool{true}, Bool{false})}

Now, it can be defined/used as:

    data Pair<A : Type, B : Type> : Type
    | new : (x : A, y : B) -> Pair

    let my_pair
        Pair<Bool, Bool>{new(Bool{true}, Bool{false})}

The main difference between parameters and indices is that
parameters are constant, i.e., don't change across different
constructors. That makes the definition of the constructors
simpler, as we don't need to receive parameters as arguments.
The main advantage, though, is that this allows discarding that
information when compiling to SIC. Indices can't be discarded
because they might change at runtime. Parameters can.

Previously, Formality couldn't compile / readback terms with
indices to SIC, which made it impossible to compile polymorphic
datatypes. Now, it still can't compile indices, but it can compile
parameters. As such, polymorphic values like the Pair above can now
be compiled to SIC, and with no runtime cost.

To compile indices to SIC, more work is needed. In particular, we
must be able to compile and readback types such as All and Idt.
Only then we'll be able to compile datatypes with indices, because,
since those indices vary during runtime, they must be encoded into
SIC programs, so we can have the full type information on readback.
Until them, datatypes like Vect won't work on SIC.

This is not very tested and there could be some bruijn indice bugs.