Formality = CoC under EAL with Self
After experimenting with even more designs, I think I've nailed it. This
one is simple to implement and use, the code is beautiful, it fully
compatible with optimal reductions and strongly normalizing. Whether it
is consistent or not is left to be seen, but I can see ways to prove it.