Unboxed 32-bit numbers and string syntax sugar
Nodes can now include unboxed 32-bit numbers, which seems to be the most
efficient way to have numeric primitives. Now, a node has 128 bits, as such:

Node (128 bits)
- port0: 32 bits
- port1: 32 bits
- port2: 32 bits
- label: 32 bits

Pointer (32 bits)
- addr: 30 bits (the global index of the node it connects to)
- slot:  2 bits (the slot it connects to: 0, 1 or 2)

Number (32 bits)
- numb: 32 bits (the number)

Label (32 bits)
- kind: 27 bits (tag of a icomb-node or operation of an op-node)
- ptyp:  3 bits (annotates if each port holds a pointer or a number)
- type:  2 bits (0 = icomb-node, 1 = op1-node, 2 = op2-node)

Numeric operations are performed with an op2 node. Its main port points
to the second number, its first port points to the first number, and its
second port points to the return location. Once it interacts with a
number, it becomes an op1 node. Its main port points to the first
number. Its first port holds the second number (it can't have a link
there). Its second port points to the return location. Once it interacts
with a number, it returns it.

Pointers are replaced by a sum type, PtrNum, which can either be an
usual pointer to a node, or a number. It needs at least 33 bits, the
additional one being used to represent what case it is. Theoretically, a
numeric PtrNum is equivalent to a pointer to a Num node, which doesn't
exist because all numbers are unboxed.

I've drawn the full system here: https://imgur.com/kmsHdUz