Pairs and If-Then-Else
If-Then-Else is necessary to allow lambdas to interact with numbers,
otherwise it would be impossible to branch based on a number's value.
Pairs make the interaction net implementation of If-Then-Else simpler,
as it fits perfectly in a 128 bit node as long as we have pairs. Native
pairs also are slightly more efficient than λ-encoded pairs, so they can
be desirable in some situations. They do not need further extensions on
the interaction net system: a pair is just a CON node with kind 0xFFFF.
One could remove pairs, but then If-Then-Else would be considerably less
elegant, and the code complexity wouldn't be improved a lot.