Inf utility for non-termination
Loops performed with Church Nats have two drawbacks:

1. They can't express non-terminating computations

2. If the step function doesn't fuse, it uses linear space

Inf is an external utility that merely applies a Formality Core term to
a value over and over, until a condition is met. It doesn't change the
language or the theory in any sense, but it does allow us to run
unbounded computations and constant-space non-fusible for loops by
defining an initial state, a step function and a stop condition.