Formality-Core to JavaScript compiler for interop
This makes it easier to interop between Formality and JavaScript. For
example, one can convert a FMC Church list to a JS array with:

fm.to_js.compile(term)(a => b => [a].concat(b))([])

An interesting note:

High-order programs which take advantage of optimality are much faster
on inets than on JS. For example, this:

(for ~10000000000 #not #true)

Is instantaneous on inets, but never finishes on JS. But this:

(for ~10000000 #[x]{x + 1} #0)

Is slightly (3x) faster with native JS functions than with inets.