Breaking syntax changes
- Lambda uses `{x y z}body` instead of `[x][y][z]body`

- Pair uses `[a,b]` instead of `&(a,b)`

- Get use `get [a,b] = v body` instead of `get &(a,b) = v body`

- Numeric operations use `|a + b|` instead of `{a + b}`

The reasoning is that numeric operations are always binary, thus don't
need open/close brackets like applications. By using `|` for them, we
free `{}`. It is, though, more usual to use `[]` for pairs, so they get
`[]`, and lambdas get `{}`. On Formality-Lang, annotated lambdas could
be written as `{a : A, b : B} => c`, and then forall could be written as
`{a : A, b : B} -> C`.