Numeric for-loop primitive
This introduces a new node type, FOR. That node has 4 ports, i.e., one
more than all others. Since this implementation has only ternary nodes,
FOR is internally represented with two nodes. Its ports hold:

- Port0: a number `n` representing the limit of the loop
- Port1: a function `f` to be applied `n` times to an argument `x`
- Port2: the argument `x` which `f` will be applied to
- Port3: the return location of the computation

                            x--|\
    ___                        | }--|\
f--| k \        (if n>0)    a--|/   |k}--f
x--|FOR |--[n]  ~~~~~~~~>        b--|/
r--|___/                        ___
                            b--|k-1\
                            a--|FOR |--[n-1]
                            r--|___/

    __
f--| k \        (if n=0)
x--|FOR |--[n]  ~~~~~~~>    r--x
r--|___/

Why is FOR necessary? It mostly isn't. You could use Church-Nats for the
same effect, in theory. Church-Nats are great because, if `f` is a
function that fuses on self-composition, they allow applying it `n` times
in `O(log(n))` space and time. But when `f` isn't, they need `O(n)`
space and time to emulate a for-loop. Time-wise, this is expected, but
for-loops shouldn't require more than constant space.

With native FOR, we're able to perform for-loops with native nums
instead of Church-Nats. It mostly works the same, but is optimized to
use constant-space on the lazy evaluator. This does NOT use
constant-space on the parallel, strict evaluator, because it unrolls the
loop too eagerly, consuming extra space. Having constant-space FOR on it
would probably require extending the system further with SEQ constructs,
allowing the programmer to ask the loop to wait until an iteration has
completed before starting the next one.