Fix bug with unary duplication
The following program:

    def list:

      let from_chars = {chars}
        let nums = {Cons}
          dup Cons = Cons
          dup func =
            let loop = {c cont char indx}
              cpy indx = indx
              cpy char = 100
              cpy done = |indx % 4|
              let CONS =
                if done
                then: (Cons 100)
                else: {x}x
              let next =
                if done
                then: 0
                else: 100
              (CONS (cont next indx))
            (chars #loop)
          # {nil} (func {char indx}nil 0 0)
        nums

      let break_lines = {list Cons}
        dup Cons = Cons
        dup fold = (list #{x xs} (Cons 1 (Cons 1 (Cons 1 xs))))
        # {Nil} (fold Nil)

      dup nil = # NilF

      # (from_chars (break_lines nil))

Caused fm-net to get into a corrup graph and loop forever. I couldn't
reproduce the problem with a smaller program. I tracker the problem to
be caused by the unary duplication rewrite, but I couldn't find the
exact cause. Re-implementing the operation based on the binary
duplication one solved the issue, at the cost of allocating two
intermediate nodes. When the binary operation is optimized to avoid this
allocation, the unary operation should be re-optimized to avoid it too.