Kaelin game design changes
All skills now use exclusively position arguments. This greatly
simplifies the UI, enabling faster turns. This means waves are removed
for now.

Walk is now a blink. This has consequences such as not being able to
bodyblock, but I actually think this may lead to a better game in many
senses. I'll be experimenting to see how it feels.