Use dependent intersections & preserve names
The idea previously was to use weak sigmas (refinement types), and then
express dependent intersections with `[x : A, ~[y : (P x), ~<x = y>]]`.
This was good because it allowed reusing the code from Sigma, but a
problem with it is that `~snd p` couldn't be erased to `p`, which
complicated things a little bit. Instead, I noticed I can just force
both elements of a "weak sigma" to be identical, and then make `~snd p`
erase to `p`... which is exactly what a dependent intersection does. So
at this point we just have dependent intersections, but I'm still
reusing the code of Sigma, which is great! Thanks Andrew for answering
my questions and sharing those insights with me.