HOAS WHNF for improved context inspection & errors
Previosuly, contexts and type errors showed fully normalized terms,
which were daunting to read. For example, an user would see:

- Found type... U32
- Instead of... (P {succ, zero} => zero)
- When checking 7
- On expression (fold 7)
- With the following context:
- n    : {~P : Type, succ : !{x : P} -> P} -> !{zero : P} -> P
- i    : {~P : {n : {~P : Type, succ : !{x : P} -> P} -> !{zero : P} -> P} -> Type, succ : !{~n : {~P : Type, succ : !{x : P} -> P} -> !{zero : P} -> P, i : (P n)} -> (P {~P, succ} => dup succ = succ; dup fold = (n ~P #succ); #{zero} => (succ (fold zero)))} -> !{zero : (P {~P, succ} => #{zero} => zero)} -> (P n)
- P    : {n : {~P : Type, succ : !{x : P} -> P} -> !{zero : P} -> P} -> Type
- succ : !{~n : {~P : Type, succ : !{x : P} -> P} -> !{zero : P} -> P, i : (P n)} -> (P {~P, succ} => dup succ = succ; dup fold = (n ~P #succ); #{zero} => (succ (fold zero)))
- succ : {~n : {~P : Type, succ : !{x : P} -> P} -> !{zero : P} -> P, i : (P n)} -> (P {~P, succ} => dup succ = succ; dup fold = (n ~P #succ); #{zero} => (succ (fold zero)))
- fold : {zero : (P {~P, succ} => #{zero} => zero)} -> (P n)
- zero : (P {~P, succ} => #{zero} => zero)

I've previously attempted to add a "show" field on each datatype, which
would remember its name, allowing it to be displayed by `show` instead
of the whole thing, but it only improved some cases.

Now, I'm instead allowing the `norm` function to reduce to the weak read
normal form, and using that to reduce types gradually on the
type-checker. This allows names (refs) to be preserved as much as
possible. With this, the same error message becomes:

Type mismatch.
- Found type... U32
- Instead of... (P {succ, zero} => zero)
- When checking 7
- On expression (fold 7)
- With the following context:
- n    : CNat
- i    : (INat n)
- P    : {n : CNat} -> Type
- succ : !{~n : CNat, i : (P n)} -> (P (CSucc n))
- succ : {~n : CNat, i : (P n)} -> (P (CSucc n))
- fold : {zero : (P CZero)} -> (P n)
- zero : (P CZero)

Which is exponentially easier to read.

(The trend of using the word "exponentially" incorrectly is growing
exponentially, so I wanted to be part of it.)