Syntax-sugar for dependent pattern-match and refactorings
This enables the following syntax:

    case<Vector> vector
    | push  => tail
    | empty => Vector.empty(~T)
    : Vector(T, Nat.pred(len))

For pattern-matching. This desugars to:

    (%vector)(
      ~ {len, self} => Vector(T, Nat.pred(len))
      , {~len, head, tail} => tail
      , Vector.empty(~T))

This also enables the following syntax:

    Nat.predif : {|n : Nat, |b : Bool} -> Nat
    | succ | true  = n.pred
           | false = Nat.succ(n.pred)
    | zero | true  = Nat.zero
           | false = Nat.zero

For dependent pattern-matching. This one is considerably more complex,
as it creates an entire patern-match tree, correctly assigning term
values on motives, and carrying resources through the branches so that
user doesn't need to duplicate them. This one desugars to:

    (%n)(
      ~{self} => {b : Bool} -> Nat
      , {n.pred, b} => (%b)(
        ~{self} => {~n : Nat, n.pred : Nat} -> Nat
        , {~n, n.pred} => n.pred
        , {~n, n.pred} => (Nat.succ)(n.pred)
        , ~(Nat.succ)(n.pred)
        , n.pred)
      , {b} => (%b)(
        ~{self} => {~n : Nat} -> Nat
        , {~n} => Nat.zero
        , {~n} => Nat.zero
        , ~Nat.zero),
      b)

Notice how `n.pred` can be used in two different branches without
copying, since it was carried to the right location with just lambdas
and applications. Notice also how an erased, reconstructed version of
`~n` is available for use inside the motives of the `b` pattern matches.
This is important when your return type depends on it (say, if we had
`{|n : Nat, |b : Bool} -> P(n,b)` instead. Finally, a lot of caution
must be taken to keep bruijn indices on motive types correct. Due to the
enourmous complexity of this, this feature should be considered very
unstable for a while, and it doesn't substitute ADT indexes yet (so, for
example, you won't be able to match a polymorphic list, since its returned
type variable will be different from the one on the type signature).

This commit also splits the language in two files, fm-core.js and
fm-lang.js. This is because the parser is getting bloated with complex
syntax sugars, so, fm-core.js will have just the simple, unsugared
language, while fm-lang.js will have the parser, stringifier and sugars.