Syntax sugar for recursive calls
Recursive functions must be annotated with `:!` and provide a "halt
case" with `*`. For example:

    Nat.double :! {n : Nat} -> Nat
      case<Nat> n
      | succ => Nat.succ(Nat.succ(Nat.double(pred)))
      | zero => Nat.zero
      : Nat
      * n

Is desugared to:

    Nat.double.call : {REC : {n : Nat} -> Nat} -> {n : Nat} -> Nat
      case<Nat> n
      | succ => Nat.succ(Nat.succ(REC(pred)))
      | zero => Nat.zero
      : Nat

    Nat.double.halt : {n : Nat} -> Nat
      n

    Nat.double
      rec(~{n : Nat} -> Nat, #Nat.double.call, #Nat.double.halt)

Note that `rec` must be in scope. Ideally it should be a very large
Church-Nat (i.e., 2^64), allowing recursive functions to have a
practically infinite bound. Note that with optimal reductions, lazy
evaluation and compact Church-Nats, this has almost no extra cost.