Recursive functions create a non-terminating copy
Recursive functions used to desugar to `rec(~{x : T} -> T, fun, halt)`,
which puts a boundary on the call count. Sometimes, though, it is
desirable to use the actual non-terminating defintion, such as when
inside an erased type. This commit causes recursive functions to create
a new definition, `-fun`, with its non-bounded version.