Syntax for binary operators and user-defined symbols
Now, mathematical operators don't need the `|a + b|` syntax. Instead,
you can write them as usual, `a + b`. Equality is `a == b`, and now you
can write `A -> B` as a shortcut for `{x : A} -> B` like on most
functional languages. Moreover, users can define their own operators by
simply making a definition that starts with one of those symbols:

    ["+","-","*","/","%","^","&","|","^","~","<",">","="]

So, for example, this is possible:

    +foo : {x : Word, y : Word} -> Word
      x + y + 1000

    main
      3 +foo 4

There is no operator precedence. The expression `2 * 3 + 4` is parsed as
`2 * (3 + 4)`. You can use parenthesis if you want `(2 * 3) + 4`.