Re-enable subset types, change not syntax
Since we don't need dependent-intersections anymore, I'm re-enabling the
`[x : A ~ B(x)]` and `[a ~ b]` syntaxes to denote subset types, i.e., to
erase the second element of a pair. This required me to avoid using the
`~` symbol on not, since it is already used for erasure. Now it uses
`-!` instead.