Rework boxed and recursive definitions, equality changes, Tak primitive
Previously, the syntax-sugar for recursive calls wasn't very flexible,
since it didn't allow 1. level-0 parameters, 2. accessing the index of
the induction, which is necessary for inductive proofs, 3. setting your
own recursion bound. This commit reworks the recursion syntax, and
affects several areas of the language.

The syntax for a boxed definition is now:

    !foo : {a : !Word} -> !{b : Word} -> Word
      3

Boxed definitions are like usual definitions, except it will
automatically unbox other boxed definitions inside it. Notice the
presence of the level-0 parameter, `a`, on the term above.

The syntax for recursive definitions now is:

    !map*n : {~A : Type, ~B : Type, f : !A -> B} -> ! {list : List(A)} -> List(B)
      case<List> list
      | cons => cons<B>(f(head), map(tail))
      | nil  => nil<B>
      : List(B)
      * nil<B>

Here, `n` is meant to represent the maximum recursion depth. In order to
call a recursive function, you must set a bound with `*`:

    !map.test : !List(Word)
        map*256<Word, Word>(#{x} x + 1, Word$[1,2,3,4])

Notice the auto-unboxing procedure lifts the entire application of `map`
to its level-0 arguments (i.e., `map*256<Word, Word>`) to the level-0.

Moreover, you're able to use `n` inside the definition, allowing you to
use the recursion syntax for inductive proofs.

This commit makes some additional changes to the language:

- There is a new primitive, Tak, which unboxes a term. The syntax for it
  is `-#(a)`. It is useful for avoiding verbose `dup`s on the type level.

- Equality checking now won't fully unbox terms. Instead, it will just
  reduce `dup x = a; b` as `b[x <- -#(a)]`. This is more flexible while
  being safer than fully unboxing. I think this will help the
  formalization of the language, for reasons I'll elaborate futurely.

- The syntax for if-then-else was changed to `if c: A else: B`.

- Definitions will automatically insert `cpy` for `Word` arguments,
  allowing you to use numeric arguments as many times as you want.

- Several minor bugfixes and adjustments.