Optimize equality checking by abusing JS's fast string comparison
This adds an extra field on all term constructors storing a compact
string representing it. Since JavaScript strings are hash consed, this
allows us to optimize considerably the equality checker by checking if
the stored strings match, and returning true prematurely if so. The
speedup obtained from this is gigantic. Future implementations of
Formality in languages without hash-consed strings should just store the
hash of the constructors to achieve the same effect.