Optimize bounded recursion, improve strat. checker
1. Optimize bounded recursion

We realized that, by using a max. recursion bound of 2^N-1, the number
of rewrites to perform a recursive function doesn't change as a function
of `N`, which is the case for other numbers. So, for example,

    add*65536(1,1)

Will need roughly 20k rewrites, while

    add*65535(1,1)

Will need roughly 1.6k rewrites only!

This is a huge difference. Because of that, now calling a recursive
function without a explicit bound will default to 2^48-1, as in:

    add*(1,1)

2. Improve strat. checker

Previously, the stratification checker wasn't capable of recognizing
certain expressions as part of the type-level. For example,

    LeftDistributive : {~A : Type, f : A -> A -> A, g : A -> A -> A} -> Type
      {x : A, y : A, z : A} -> f(x, g(y,z)) == g(f(x,y),f(x,z))

    RightDistributive : {~A : Type, f : A -> A -> A, g : A -> A -> A} -> Type
      {x : A, y : A, z : A} -> f(g(y,z), x) == g(f(y,x),f(z,x))

    Distributive : {~A : Type, f : A -> A -> A, g : A -> A -> A } -> Type
      Conjunction(LeftDistributive(~A,f,g), RightDistributive(~A,f,g))

Here, `Distributive` returns a type, so the variables `f` and `g`
shouldn't be restricted by stratification checks, but is. This commit
adds a new primitive, `type(t)`, which performs a type-level identity.
This primitive allows the stratification checker to be disabled whenever
you have an expression that returns a type.

Moreover, the annotated function syntax will wrap the function around a
`type(...)` when it returns a Type. As such, the definition above works
with no change.