Rework import system
The old import system grew up too chaotically, which led to some bugs,
unclean code and inefficiency. This reworks the system to flatten the
import tree, such that there is only one `def` for each import, and
qualified/open refs are resolved on the spot when constructing a Ref.
This solves important bugs and makes the system considerably more
efficient. Usage should stay roughly the same.