New case-tree syntax, revert parse-infer and @ case-syntax
This commit changes again the case syntax from:

  @Foo x ~> T
  | ctr0 => body0
  | ctr1 => body1

Back to:

  case/Foo x
  | ctr0 => body0
  | ctr1 => body1
  : T

The reason is that having partial inference on the parser was more
complex than I was comfortable with, and only worked for a very limited
amount of cases, so it felt inconsistent. Moreover, in retrospect, the
symbolic syntax isn't as clean as I wish, and the case syntax is more
obvious / least surprising.

Note that a `/` is now used instead of `<>` as that is less misleading
since it expects a name, not a type. That is, it feels inconsistent to
have `new<List(A)>`, but `case<List>`. The `/` syntax will be used for
future sugars such as `do/MyMonad` and `open/MyRecord x`.

Now, in order to reduce pattern-matching noise, I've took a different
approach and re-implemented case'd args using a different approach. This
will correctly deal with ADT parameters, but isn't capable of dealing
with indexed datatypes. This is fine, though: if an user is advanced
enough to use indexed datatypes, he/she is probably able to understand
the full case syntax anyway!

The case'd syntax requires you to annotate a top-level function argument
with the `case` keyword, and then list all the cases in the lines below
it. It will automatically pass function arguments down the case-of
branches, avoiding duplications, so you can "use" a function variable
twice, as long as it is in different branches.

This is considerably cleaner. For example, `add` goes from:

    !add*N : !{n : Nat, *m : Nat} -> Nat
      (case/Nat n
      | succ => {m} succ(add(pred, m))
      | zero => {m} m
      : Nat -> Nat)(m)

to:

    !add*N : !{case n : Nat, *m : Nat} -> Nat
    | succ => succ(add(n.pred, m))
    | zero => m

Functions with multiple cases are specially favored by not having to
manually pass arguments down the branches. `or` goes from:

    or : {b0 : Bool, b1 : Bool} -> Bool
      (@ b0      ~> Bool -> Bool
      | true     => {b1} @Bool b1 ~> Bool
        | true   => true
        | false  => true
      | false    => {b1} @Bool b1 ~> Bool
        | true   => true
        | false  => false)(b1)

to:

    or : {case b0 : Bool, case b1 : Bool} -> Bool
    | true  true  => true
    | true  false => true
    | false true  => true
    | false false => false

The sugar will also fill `self`s and opened values in motives. For
example, `demorgan`'s theorem goes from:

    demorgan1b : {a : Bool, b : Bool} -> not(and(a,b)) == or(not(a), not(b))
      (case/Bool a
      | true    => {b}
        case/Bool b
        | true  => refl<false>
        | false => refl<true>
        : not(and(true, self)) == or(not(true), not(self))
      | false   => {b}
        case/Bool b
        | true  => refl<true>
        | false => refl<true>
        : not(and(false, self)) == or(not(false), not(self))
      : {b : Bool} -> not(and(self, b)) == or(not(self), not(b)))(b)

To merely:

    demorgan0 : {case a : Bool, case b : Bool} -> not(or(a,b)) == and(not(a), not(b))
    | true  true  => refl<false>
    | true  false => refl<false>
    | false true  => refl<false>
    | false false => refl<true>