Imports open by default, improve name resolution
Now all imports are open by default. If there is an ambiguity, such as
attempting to use a `foo` that was defined in two different imports, the
parser will ask you to desambiguate. It is OK for two imports to define
the same name, as long as you don't use that name unqualified. Since
file names are globally unique, a full name (File@V/term) will never
conflict.