Floating point operations, unary operators, consistent operator parsing
The following operators were added:

    +f  float addition
    -f  float subtraction
    *f  float multiplication
    /f  float division
    %f  float modulus
    ^f  float exponentiation
    .f  converts an integer to a float
    .u  converts a float to an integer

Note that floats are represented on Words, there is no new numeric type.
Those operations merely view the bits in a Word as encoding a Float.

The `**` operator was removed, since it is obsolete now.

The `.!`, `.f` and `.u` operators can be used unarily, as in `.!(x)`.

The parsing for all operator-like features now use the same parse logic for
consistency. Note operators can't spam multiple lines. This includes,
for example:

    Applications   f(x)
    Lists          Word $ [1, 2, 3]
    Rewrite        foo :: rewrite x as P(x) with e
    Operators      3 + 7