Compact syntax for nested sigmas
Nested sigmas can now be written as:

    [x : A, y : P(x), z : Q(x, y), R(x, y, z)]

And nested pairs can be written as:

    [1, 2, 3, 4]

When the second element of a pair is erased, it doesn't need a `,`
anymore. That's to avoid an ambiguity. So, for example:

    [x : A, y : B ~ z : C, D]

Is parsed as the old:

    [x : A, [y : B, ~[z : C, D]]]

And:

    [x : A, y : B, ~z : C, D]

Is parsed as the old:

    [x : A, [y : B, [~z : C, D]]]

And:

    [x : A, y : B ~ ~z : C, D]

Is parsed as the old:

    [x : A, [y : B, ~[~z : C, D]]]