Type logging and typed holes.
Typed holes will now display an error with the expected type and
context, and will be assumed to be true by the type-checker, allowing
you to implement it later. When using Formality as a library, you can
disable this by calling `typecheck` with `debug = false`.

The `log` primitive can now be used to print types on type-checking.
This is very useful for inspecting the type of variables that aren't on
context. For example, the term:

    main : {to_bool : {u : Unit} -> Bool, u : Unit} -> Bool
      let b = to_bool(u)
      ?help

Displays the following message:

    [ERROR]
    Hole found: 'help'.
    - With goal... Bool
    - Inside of... {u} => ?help
    - With context:
    - to_bool : {u : Unit} -> Bool
    - u       : Unit

Notice it doesn't show the type of `b`, since `let` performs parse-time
substitutions. You can inspect it with `log`:

    main : {to_bool : {u : Unit} -> Bool, u : Unit} -> Bool
      let b = to_bool(u)
      log(b)
      ?

This will output:

    [LOG]
    Term: to_bool(u)
    Type: Bool

Informing us that `b` has the type needed by our hole.