Change halt case to `halt:` instead of `*`
The `*` was way too overloaded: it was used for word multiplication, to
introduce the call count of a recursive definition, to set the call
limit of a recursive function, and to set the halt case. The later was
different from the rest, and is not "obvious", so I think a keyword here
makes it more likely for someone to understand what is going on from the
first impression, and improves readability. Examples:

    #add*N : !{case n : Nat, halt m : Nat} -> Nat
    | succ => succ(add(n.pred, m))
    | zero => m

    #to_word*N : !{case n : Nat} -> Word
    | succ => 1 + to_word(n.pred)
    | zero => 0
    halt: 0