Improve bidirectional type-checker of 'get'
Previously, since pairs are sigmas, the type-checker couldn't carry the
expected type of the `get [a,b] = pair; body` expression to its body.
That's because the type of body could depend on `a` and `b`. Because of
that, something like the program below:

    main : {pair : [x : Word, x == x]} -> [x : Word, x == x]
      get [a, e] = pair
      [a, e]

Wouldn't type-check, because `[a, e]` is inferred to have the type
`[x : Word, a == a]`. Note that `a` refers to the variable bound by
`get`, not by the `x` of the sigma. Outside of `get`, that will become
`[x : Word, fst(pair) == fst(pair)]`, which is different from the type
expected, `[x : Word, x == x]`. To solve that, you could add an explicit
type-annotation to the pair:

    main : {pair : [x : Word, x == x]} -> [x : Word, x == x]
      get [a, e] = pair
      [a, e] :: [x : Word, x == x]

This would type-check, because annotated sigmas can use the information
that `a == x`, and, thus, return a second element that depends on `x`,
not on the `a` bound by the `get`. In general, dependent pairs must be
annotated, because inferring the type of a simple pair is limited.

A possible solution would be to extend the syntax of simple pairs to
allow annotations, as in, `[a, e : P(a)]`, in the same way we can
annotate the argument of lambdas, but I'll not do that for now, since it
complicates the syntax of the language.

Instead, I'll slightly improve the behavior of `get` as follows. Now, it
will propagate the expected type to its body by simply ignoring the two
variables bound by it. In most cases, such as the example above, this
will remove the need of explicitly annotating its body, and everything
will work as expected. This has a problem, though: if the body does
depend on the bound variables, then it will be impossible to type-check
it, since the expected type now ignores them.  To solve that, if we get
a type-error on the first attempt, we'll try again, this time without an
expected type, falling back to the old behavior.