Smarter case expressions, cong primitive, move show back to fm-lang.js
Formality now has a `cong` primitive, `cong(~f, ~e)`, which, given a
proof of type `e : a == b`, and a `f`, returns `e : f(a) == f(b)`.

The `show` functioin was moved back to fm-lang.js.

The syntax of case expressions received new several improvements.

1. Fields are accessed with `name.field`

Previously, cased arguments used `name.field`, while case expressions
used just `field`. Now, both use `name.field`. When possible, Formality
will use the name of the variable, as in:

    case/And x
    | both => x.left + x.right
    : Word

When not possible, you may provide a name with `as`:

    case/And f(a) as x
    | both => x.left + x.right
    : Word

2. The "self" keyword was replaced by the matched name

As an example, this:

  case/Bool a
  | true  => refl(~true)
  | false => refl(~false)
  : not(not(self)) == self

Is now written as:

  case/Bool a
  | true  => refl(~true)
  | false => refl(~false)
  : not(not(a)) == a

To access the upper 'a', you must use 'a^'.

3. New feature: resource moving

Now, the `move t : T as x` expression can be used before the cases to
move the term `t` of type `T` to the branches with the name `x`. When
`t` is a variable, the `as x` is implicit. For example, this program:

    xor : {a : Bool, b : Bool} -> Bool
      (case/Bool a
      | true  => {b} not(b)
      | false => {b} b
      : {x : Bool} -> Bool)(b)

Can now be written as:

    main : {a : Bool, b : Bool} -> Bool
      case/Bool a
      move b : Bool
      | true  => not(b)
      | false => b
      : Bool

This considerably improves the readability of motives and branches,
specially when many resources with large types are passed down.

4. New feature: free equality proofs

Now, when you pattern-match against a structure, Formality will
automatically add an equality proof specializing the matched value to
its shape. For example, when matching against a Bool `x`, you will get a
`x.equal : x == true` proof on the true case, and a `x.equal x == false`
proof on the false case. That desugars to an application of `refl`, and
allows simplifying proofs. Examples:

This:

    (case/Bool true as x
    | true  => {~e} 10
    | false => {~e} absurd(true_isnt_false(x.equal), ~Word)
    : {~e : true == x} -> Word)(~refl(~true))

Can now be written as:

    main : Word
      case/Bool true as x
      | true  => 10
      | false => absurd(true_isnt_false(x.equal), ~Word)
      : Word

You can also get proofs of indices using the `with x.idx == t`
expression. Examples:

This:

    vhead : {~T : Type, ~n : Nat, xs : Vector(T, succ(n))} -> T
      (case/Vector xs
      | vcons => {~e} xs.head
      | vnil  => {~e} absurd(succ_isnt_zero(~n, ~sym(~e)), ~T)
      : {~e : xs.len == succ(n)} -> T
      )(~refl(~succ(n)))

Can now be written as:

    vhead : {~T : Type, ~n : Nat, xs : Vector(T, succ(n))} -> T
      case/Vector xs
      with xs.len == succ(n)
      | vcons => xs.head
      | vnil  => absurd(succ_isnt_zero(~n, ~xs.len.equal), ~T)
      : T

This considerably improves the readability and usability of case
expressions when you need to exploit index equalities.