Add primitive for eta-reduction
Previously, it wasn't possible to prove that `[fst(a),snd(a)] == a` for
native pairs, nor to prove that `{x} f(x) == f` for native functions.
This also meant that we couldn't have, for example, dependent
elimination on native pairs, even thought it was possible for an
user-defined pair. Now, `eta(~t)` can be used to access the uniqueness
principle of lambdas and pairs as follows:

eta_expand_lambda : {A : Type, B : {x : A} -> Type, fn : {x : A} -> B(x)} -> fn == {x : A} fn(x)
  eta(~fn)

eta_expand_pair : {A : Type, B : Type, ab : [:A,B]} -> ab == [fst(ab),snd(ab)]
  eta(~ab)