Add fm-json module for JS-Formality interop
Before, when calling formality a function from JS, one would need
to manually create a Formality-Core term for the argument and then
create another term by applying the argument to the function.

This commit eases the creation of the argument term by introducing
a special type that represents a Json in formality and then creates
translation functions to make it easier to generate that.

The code for the Json type isn't on Base now, but any types that
has the same structure as the following will work

    T JsNumber
    | js_number { value : Word }

    T Json
    | null
    | float32 {value: JsNumber}
    | uint32 {value: Word}
    | string {value: String}
    | list {value : List(Json)}
    | object {value : List([:String, Json])}

Now, let's say there is a function Module/term of type Json -> Json.
In order to call from Javascript, one should do:

    // convert a value from JS to a Formality Term
    const input_term = json.to(input);

    // Creates a new term applying the value to a function
    const applied_term = core.App(core.Ref("Module/term"), input_term, false)

    // Reduces the new term
    const net = to_net.compile(lang.erase(applied_term, defs), defs)
    net.reduce_strict({})
    const output_term = to_net.decompile(net)

    // convert the resulting term to an actual JS value
    const output = json.from(output_term)

Is not the objective of this commit to ease the application term
construction, that is left to a future commit.