Check recursive calls on rewrite evidence
Otherwise, one could derive an inconsistency:

    bad : true == false
      unit :: rewrite x
        in case/Bool x
          | true  => Unit
          | false => true == false
          : Type
        with bad

The evidence of the equality is relevant. Cedille's equality erases the
evidence, though, which is still confusing to me. This requires more
attention, and I highly suspect that this equality is still problematic.
We can encode equality with self-types anyway, so I'm not sure having
equality primitives is desirable.