Prevent erased arguments from being un-erased
Previously, it was possible to use erased arguments in computational
positions, allowing us to prove Empty in different ways:

    bad : {P : Type} -> P
      bad(P)

    id : {~A : Type, ~a : A} -> A
      a

    tif : {~e : true == false} -> Empty
      true_isnt_false(e)

    main : Empty
      tif(~bad(true == false))

    foo : Empty
      id(~Empty, ~bar)

    bar : Empty
      id(~Empty, ~foo)

This commit fixes this issue.

It is important to note that, since we allow unrestricted recursion on
erased arguments, they can't be interpreted as meaningfully affecting a
function's domain. For example, this `main` normalizes to 7:

    bad : {P : Type} -> P
      bad(P)

    aff : {x : Word, ~e : x == 0} -> Word
      x

    main
      aff(7, ~bad(7 == 0))

Even though one would expect it to be only zero. The question of how to
have inductive datatypes like Ind and Vector on Formality is a though
one. While we can have recursion on types such as Scott Nats and Lists,
inductive datatypes like Ind and Vector require some amount of recursion
and unboxing on their erased arguments, yet doing so in an unrestricted
manner can lead to inconsistencies.

A possible solution is to have two types of erased arguments: one for
arguments used on types and other for computational irrelevance. Before
picking to a particular solution, though, Formalization and consistency
proof of fm-core should be priority, and we should refrain from building
too many base libraries and language features before, as those may end
up relying on potentially unsound features of fm-core.